perf: cache crv APY dict
