+++ src/cached.ts
@@ -1,13 +1,15 @@
 import memoize from "memoizee";
 import {IDict, IExtendedPoolDataFromApi, INetworkName, IPoolType} from "./interfaces.js";
-import {uncached_getAllPoolsFromApi, uncached_getUsdPricesFromApi} from './external-api.js'
+import {uncached_getAllPoolsFromApi, uncached_getCrvApyFromApi, uncached_getUsdPricesFromApi} from './external-api.js'
+import {curve} from "./curve";
 
 const _getCachedData = memoize(
     async (network: INetworkName, isLiteChain: boolean) => {
-        const allPools = await uncached_getAllPoolsFromApi(network, isLiteChain);
-        const poolLists = Object.values(allPools)
+        const poolsDict = await uncached_getAllPoolsFromApi(network, isLiteChain);
+        const poolLists = Object.values(poolsDict)
         const usdPrices = uncached_getUsdPricesFromApi(poolLists);
-        return { allPools, poolLists, usdPrices }
+        const crvApy = uncached_getCrvApyFromApi(poolLists)
+        return { poolsDict, poolLists, usdPrices, crvApy };
     },
     {
         promise: true,
@@ -17,8 +19,8 @@ const _getCachedData = memoize(
 
 export const _getPoolsFromApi =
     async (network: INetworkName, poolType: IPoolType, isLiteChain = false): Promise<IExtendedPoolDataFromApi> => {
-        const {allPools} = await _getCachedData(network, isLiteChain);
-        return allPools[poolType]
+        const {poolsDict} = await _getCachedData(network, isLiteChain);
+        return poolsDict[poolType]
     }
 
 export const _getAllPoolsFromApi = async (network: INetworkName, isLiteChain = false): Promise<IExtendedPoolDataFromApi[]> => {
@@ -27,6 +29,13 @@ export const _getAllPoolsFromApi = async (network: INetworkName, isLiteChain = f
 }
 
 export const _getUsdPricesFromApi = async (): Promise<IDict<number>> => {
-    const {usdPrices} = await _getCachedData("ethereum", false);
+    const network = curve.constants.NETWORK_NAME;
+    const {usdPrices} = await _getCachedData(network, false);
     return usdPrices
+}
+
+export const _getCrvApyFromApi = async (): Promise<IDict<[number, number]>> => {
+    const network = curve.constants.NETWORK_NAME;
+    const {crvApy} = await _getCachedData(network, false);
+    return crvApy
 }
